fix index data type

+++ arctic/tickstore/tickstore.py
@@ -866,7 +866,7 @@ class TickStore(object):
         rtn[END] = end
         rtn[START] = start
         if isinstance(data['index'][0], np.datetime64):
-            idx = np.concatenate(([data['index'][0]], np.diff(data['index'])))
+            idx = np.concatenate(([data['index'][0].astype(np.uint64)], np.diff(data['index']).astype(np.uint64)))
             # causal conversion ns -> ms
             if index_precision == 's':
                 s = 1_000_000_000
@@ -876,7 +876,7 @@ class TickStore(object):
                 raise ValueError(index_precision)
             idx = np.floor_divide(idx + np.uint64(s-1), np.uint64(s)).astype(np.uint64)
         else:
-            idx = np.concatenate(([data['index'][0]], np.diff(data['index'])))
+            idx = np.concatenate(([data['index'][0].astype(np.uint64)], np.diff(data['index']).astype(np.uint64)))
             if index_precision == 's':
                 idx = np.floor_divide(idx + np.uint64(1_000 - 1), np.uint64(1_000)).astype(np.uint64)
                  # idx = -(-idx // 1000)  # ceiling int div (alternative) TODO doesnt work with uint*
